# target/swagger.json
{"swagger":"2.0","info":{"version":"1.0.0","title":"piweb: Raspberry PI REST Web API"},"basePath":"/piweb/v1","tags":[{"name":"observers"},{"name":"descriptions"},{"name":"ports"}],"schemes":["http"],"paths":{"/descriptions":{"get":{"tags":["descriptions"],"summary":"List configuration of all ports","description":"","operationId":"getPortConCollection","produces":["application/json"],"parameters":[],"responses":{"200":{"description":"successful operation","schema":{"$ref":"#/definitions/PortDescriptionList"}}}}},"/observers":{"get":{"tags":["observers"],"summary":"List all registered port obeservers","description":"Returns status of single port","operationId":"getAllObservers","produces":["application/json"],"parameters":[],"responses":{"200":{"description":"successful operation","schema":{"$ref":"#/definitions/RaspberryPort"}}}},"put":{"tags":["observers"],"summary":"Register observer","description":"Registers a new observer","operationId":"putNewObserver","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","description":"New observer","required":false,"schema":{"$ref":"#/definitions/PortObserver"}}],"responses":{"200":{"description":"successful operation","schema":{"$ref":"#/definitions/PortObserver"}}}}},"/observers/{id}":{"get":{"tags":["observers"],"summary":"Find observer by observer-ID","description":"Returns info of single observer","operationId":"getObserver","produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"ID of observer","required":true,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"$ref":"#/definitions/PortObserver"}}}},"delete":{"tags":["observers"],"summary":"Unregister observer","description":"Removes the observer with the given ID","operationId":"deleteObserver","parameters":[{"name":"id","in":"path","description":"ID of observer","required":true,"type":"string"}],"responses":{"default":{"description":"successful operation"}}}},"/ports":{"get":{"tags":["ports"],"summary":"List all ports","description":"Returns status of all ports","operationId":"getAllPortStatus","produces":["application/json"],"parameters":[],"responses":{"200":{"description":"successful operation","schema":{"type":"array","items":{"type":"object"}}}}}},"/ports/{id}":{"get":{"tags":["ports"],"summary":"Find port ID","description":"Returns status of single port","operationId":"getPortStatus","produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"ID of port","required":true,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"$ref":"#/definitions/RaspberryPort"}}}},"post":{"tags":["ports"],"summary":"Set port status of specific port to 0/1","description":"","operationId":"setPortStatus","produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"ID of port","required":true,"type":"string"},{"in":"body","name":"body","description":"New state of port ('0', '1', 'false','true')","required":false,"schema":{"type":"string"}}],"responses":{"201":{"description":"Success: port status set"},"400":{"description":"No output port"},"500":{"description":"Invalid ID supplied"}}}}},"definitions":{"PortObserver":{"type":"object","properties":{"observerId":{"type":"string"},"portId":{"type":"string"},"observerURL":{"type":"string"}}},"RaspberryPort":{"type":"object","properties":{"id":{"type":"string"},"value":{"type":"boolean","default":false},"isOutput":{"type":"boolean","default":false}}},"PortDescriptionList":{"type":"object","properties":{"portDescriptions":{"type":"array","items":{"$ref":"#/definitions/PortDescription"}}}},"PortDescription":{"type":"object","properties":{"raspPortNr":{"type":"integer","format":"int32"},"id":{"type":"string"},"description":{"type":"string"},"isOutput":{"type":"boolean","default":false}}}}}